Now using standard request code for `PHOTO_EDITOR_REQUEST`
The current request code simply seems to be the timestamp when it was done, but it needs to be a 16 bit int. See #21 and https://developer.android.com/training/basics/intents/result

It doesn't have to be unique across apps, it's just to validate that the results comes from that request. It could be any other 16 bit number if desired.

Fixes #21